Move wrap to lib to allow editable to use it
